refactor(test): migrate test_summarizer unique tests to test_ai_summarizer
迁移 2 个唯一测试（discussion 链接省略和中文标签）至 test_ai_summarizer.py，删除旧文件
